--- tests/kernel/obj_tracing/src/phil.h
@@ -1,8 +1,0 @@
-/* phil.h - dining philosophers header file*/
-
-/*
- * Copyright (c) 2011-2016 Wind River Systems, Inc.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-#define N_PHILOSOPHERS  5
